# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "VocaDbWeb",
    "version": "1.0"
  },
  "paths": {
...
    "/api/releaseEventSeries/{id}": {
      "delete": {
        "tags": [
          "ReleaseEventSeriesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "notes",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "hardDelete",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "ReleaseEventSeriesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ReleaseEventSeriesOptionalFields"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentLanguagePreference"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForApiContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForApiContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForApiContract"
                }
              }
            }
          }
        }
      }
    },
    "/api/releaseEventSeries": {
      "get": {
        "tags": [
          "ReleaseEventSeriesApi"
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "fields",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ReleaseEventSeriesOptionalFields"
            }
          },
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "maxResults",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "getTotalCount",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "nameMatchMode",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NameMatchMode"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ContentLanguagePreference"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForApiContractPartialFindResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForApiContractPartialFindResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForApiContractPartialFindResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/releaseEventSeries/{id}/for-edit": {
      "get": {
        "tags": [
          "ReleaseEventSeriesApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForEditForApiContract"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForEditForApiContract"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseEventSeriesForEditForApiContract"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistForEventContract": {
        "type": "object",
        "properties": {
          "artist": {
            "$ref": "#/components/schemas/ArtistContract"
          },
          "effectiveRoles": {
            "$ref": "#/components/schemas/ArtistEventRoles"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "roles": {
            "$ref": "#/components/schemas/ArtistEventRoles"
          }
        },
        "additionalProperties": false
      },
...
      "ContentLanguagePreference": {
        "enum": [
          "Default",
          "Japanese",
          "Romaji",
          "English"
        ],
        "type": "string"
      },
      "ContentLanguageSelection": {
        "enum": [
          "Unspecified",
          "Japanese",
          "Romaji",
          "English"
        ],
        "type": "string"
      },
...
      "EntryStatus": {
        "enum": [
          "Draft",
          "Finished",
          "Approved",
          "Locked"
        ],
        "type": "string"
      },
      "EntryThumbForApiContract": {
        "type": "object",
        "properties": {
          "mime": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "urlOriginal": {
            "type": "string",
            "nullable": true
          },
          "urlSmallThumb": {
            "type": "string",
            "nullable": true
          },
          "urlThumb": {
            "type": "string",
            "nullable": true
          },
          "urlTinyThumb": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EventCategory": {
        "enum": [
          "Unspecified",
          "AlbumRelease",
          "Anniversary",
          "Club",
          "Concert",
          "Contest",
          "Convention",
          "Other",
          "Festival"
        ],
        "type": "string"
      },
...
      "LocalizedStringContract": {
        "type": "object",
        "properties": {
          "language": {
            "$ref": "#/components/schemas/ContentLanguageSelection"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LocalizedStringWithIdContract": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "language": {
            "$ref": "#/components/schemas/ContentLanguageSelection"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NameMatchMode": {
        "enum": [
          "Auto",
          "Partial",
          "StartsWith",
          "Exact",
          "Words"
        ],
        "type": "string"
      },
...
      "PVContract": {
        "type": "object",
        "properties": {
          "author": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          },
          "extendedMetadata": {
            "$ref": "#/components/schemas/PVExtendedMetadata"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "length": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "publishDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pvId": {
            "type": "string",
            "nullable": true
          },
          "service": {
            "$ref": "#/components/schemas/PVService"
          },
          "pvType": {
            "$ref": "#/components/schemas/PVType"
          },
          "thumbUrl": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PVExtendedMetadata": {
        "type": "object",
        "properties": {
          "json": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReleaseEventForApiContract": {
        "type": "object",
        "properties": {
          "additionalNames": {
            "type": "string",
            "nullable": true
          },
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistForEventContract"
            },
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mainPicture": {
            "$ref": "#/components/schemas/EntryThumbForApiContract"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalizedStringContract"
            },
            "nullable": true
          },
          "pvs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PVContract"
            },
            "nullable": true
          },
          "series": {
            "$ref": "#/components/schemas/ReleaseEventSeriesContract"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "seriesNumber": {
            "type": "integer",
            "format": "int32"
          },
          "seriesSuffix": {
            "type": "string",
            "nullable": true
          },
          "songList": {
            "$ref": "#/components/schemas/SongListBaseContract"
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagUsageForApiContract"
            },
            "nullable": true
          },
          "urlSlug": {
            "type": "string",
            "nullable": true
          },
          "venue": {
            "$ref": "#/components/schemas/VenueForApiContract"
          },
          "venueName": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "webLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebLinkForApiContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ReleaseEventSeriesContract": {
        "type": "object",
        "properties": {
          "additionalNames": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "deleted": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "pictureMime": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "urlSlug": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "webLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebLinkContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReleaseEventSeriesForApiContract": {
        "type": "object",
        "properties": {
          "additionalNames": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseEventForApiContract"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mainPicture": {
            "$ref": "#/components/schemas/EntryThumbForApiContract"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalizedStringContract"
            },
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "urlSlug": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "webLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebLinkForApiContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReleaseEventSeriesForApiContractPartialFindResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseEventSeriesForApiContract"
            },
            "nullable": true
          },
          "term": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ReleaseEventSeriesForEditForApiContract": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/EventCategory"
          },
          "defaultNameLanguage": {
            "$ref": "#/components/schemas/ContentLanguageSelection"
          },
          "deleted": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "mainPicture": {
            "$ref": "#/components/schemas/EntryThumbForApiContract"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalizedStringWithIdContract"
            },
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "webLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebLinkForApiContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReleaseEventSeriesOptionalFields": {
        "enum": [
          "None",
          "AdditionalNames",
          "Description",
          "Events",
          "MainPicture",
          "Names",
          "WebLinks"
        ],
        "type": "string"
      },
...
      "SongListBaseContract": {
        "type": "object",
        "properties": {
          "featuredCategory": {
            "$ref": "#/components/schemas/SongListFeaturedCategory"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TagUsageForApiContract": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "tag": {
            "$ref": "#/components/schemas/TagBaseContract"
          }
        },
        "additionalProperties": false
      },
...
      "VenueForApiContract": {
        "type": "object",
        "properties": {
          "additionalNames": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "addressCountryCode": {
            "type": "string",
            "nullable": true
          },
          "coordinates": {
            "$ref": "#/components/schemas/OptionalGeoPointContract"
          },
          "deleted": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseEventContract"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "names": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalizedStringContract"
            },
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "webLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebLinkForApiContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WebLinkCategory": {
        "enum": [
          "Official",
          "Commercial",
          "Reference",
          "Other"
        ],
        "type": "string"
      },
...
      "WebLinkForApiContract": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/WebLinkCategory"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "descriptionOrUrl": {
            "type": "string",
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}